created separate fronte end service request to show all reports independent of user
