+++ src/utilities/reports-api.js
@@ -1,22 +1,10 @@
 import sendRequest from "./send-request";
-import BASE_URL from "/api/reports";
-
-export async function create(report) {
-  return sendRequest(`${BASE_URL}/account`, "POST", report);
-}
+const BASE_URL = "/api/reports";
 
 export async function getAllReports() {
-  return sendRequest(`${BASE_URL}/account`);
+  return sendRequest(`${BASE_URL}/`);
 }
 
 export async function getOneReport(id) {
-  return sendRequest(`${BASE_URL}/account/${id}`);
-}
-
-export async function editReport(id, report) {
-  return sendRequest(`${BASE_URL}/account/${id}`, "PUT", report);
-}
-
-export async function deleteReport(id) {
-  return sendRequest(`${BASE_URL}/account/${id}`, "DELETE");
+  return sendRequest(`${BASE_URL}/${id}`);
 }
