+++ src/utilities/reports-service.js
@@ -1,21 +1,9 @@
 import * as reportsAPI from "./reports-api";
 
-export async function addReport(report) {
-  return reportsAPI.create(report);
-}
-
-export async function getReports() {
+export async function getAllReports() {
   return reportsAPI.getAllReports();
 }
 
-export async function getReport(id) {
+export async function getOneReport(id) {
   return reportsAPI.getOneReport(id);
-}
-
-export async function updateReport(id, report) {
-  return reportsAPI.editReport(id, report);
-}
-
-export async function deleteReport(id) {
-  return reportsAPI.deleteReport(id);
 }
